Article_Collector/requirements.txt
fastapi
uvicorn
python-multipart
asyncio

requests
beautifulsoup4
lxml
playwright
Crawl4AI==0.6.3

spacy
# torch>=2.6  # 기본 Docker 이미지의 버전을 사용하기 위해 주석 처리
transformers>=4.53.0
sentencepiece==0.2.0
nltk
openai
google-cloud-storage
google-generativeai
SQLAlchemy
PyMySQL

protobuf
pyyaml
aiohttp
cchardet
aiodns
langdetect
pytrends
python-slugify
gunicorn
https://github.com/explosion/spacy-models/releases/download/ko_core_news_sm-3.8.0/ko_core_news_sm-3.8.0-py3-none-any.whl#sha256=15486b93a22ec8ee43f787390afdb4e2086851573eb12b69dd2c4f2a95b3e0a4
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
konlpy==0.6.0
kiwipiepy==0.16.2
kss==4.5.4
mecab-python
accelerate
bitsandbytes
scikit-learn
newspaper3k
lxml